+++ sqlalchemy_utils/__init__.py
@@ -7,7 +7,6 @@ from .asserts import (  # noqa
     assert_nullable
 )
 from .batch import batch_fetch, with_backrefs  # noqa
-from .decorators import generates  # noqa
 from .exceptions import ImproperlyConfigured  # noqa
 from .expression_parser import ExpressionParser  # noqa
 from .functions import (  # noqa